Attempt to use ip4 fallback in OwnCloudClient too
