+++ src/main/java/com/owncloud/android/lib/common/OwnCloudClient.java
@@ -1,23 +1,23 @@
 /* ownCloud Android Library is available under MIT license
  *   Copyright (C) 2015 ownCloud Inc.
  *   Copyright (C) 2012  Bartek Przybylski
- *   
+ *
  *   Permission is hereby granted, free of charge, to any person obtaining a copy
  *   of this software and associated documentation files (the "Software"), to deal
  *   in the Software without restriction, including without limitation the rights
  *   to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  *   copies of the Software, and to permit persons to whom the Software is
  *   furnished to do so, subject to the following conditions:
- *   
+ *
  *   The above copyright notice and this permission notice shall be included in
  *   all copies or substantial portions of the Software.
- *   
- *   THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, 
+ *
+ *   THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
  *   EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
- *   MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND 
- *   NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS 
- *   BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN 
- *   ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN 
+ *   MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ *   NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS
+ *   BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN
+ *   ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
  *   CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  *   THE SOFTWARE.
  *
@@ -27,6 +27,7 @@ package com.owncloud.android.lib.common;
 
 import android.net.Uri;
 
+import com.nextcloud.common.DNSCache;
 import com.owncloud.android.lib.common.accounts.AccountUtils;
 import com.owncloud.android.lib.common.network.RedirectionPath;
 import com.owncloud.android.lib.common.network.WebdavUtils;
@@ -48,13 +49,14 @@ import org.apache.commons.httpclient.params.HttpParams;
 
 import java.io.IOException;
 import java.io.InputStream;
+import java.net.SocketTimeoutException;
 import java.util.Locale;
 
 import lombok.Getter;
 import lombok.Setter;
 
 public class OwnCloudClient extends HttpClient {
-	
+
     private static final String TAG = OwnCloudClient.class.getSimpleName();
     public static final int MAX_REDIRECTIONS_COUNT = 3;
     private static final String PARAM_SINGLE_COOKIE_HEADER = "http.protocol.single-cookie-header";
@@ -67,12 +69,12 @@ public class OwnCloudClient extends HttpClient {
 
 
     private static byte[] sExhaustBuffer = new byte[1024];
-    
+
     private static int sIntanceCounter = 0;
     @Getter @Setter private boolean followRedirects = true;
     @Getter private OwnCloudCredentials credentials = null;
     private int mInstanceNumber;
-    
+
     @Getter private Uri baseUri;
     @Setter private String userId;
 
@@ -81,31 +83,31 @@ public class OwnCloudClient extends HttpClient {
      */
     public OwnCloudClient(Uri baseUri, HttpConnectionManager connectionMgr) {
         super(connectionMgr);
-        
+
         if (baseUri == null) {
         	throw new IllegalArgumentException("Parameter 'baseUri' cannot be NULL");
         }
         this.baseUri = baseUri;
-        
+
         mInstanceNumber = sIntanceCounter++;
         Log_OC.d(TAG + " #" + mInstanceNumber, "Creating OwnCloudClient");
 
         String userAgent;
 
         userAgent = OwnCloudClientManagerFactory.getUserAgent();
-        
+
         getParams().setParameter(HttpMethodParams.USER_AGENT, userAgent);
         getParams().setParameter(PARAM_PROTOCOL_VERSION, HttpVersion.HTTP_1_1);
         getParams().setCookiePolicy(CookiePolicy.IGNORE_COOKIES);
         // to avoid problems with some web servers
-        getParams().setParameter(PARAM_SINGLE_COOKIE_HEADER, PARAM_SINGLE_COOKIE_HEADER_VALUE); 
-        
+        getParams().setParameter(PARAM_SINGLE_COOKIE_HEADER, PARAM_SINGLE_COOKIE_HEADER_VALUE);
+
         applyProxySettings();
-        
+
         clearCredentials();
     }
 
-    
+
     private void applyProxySettings() {
     	String proxyHost = System.getProperty("http.proxyHost");
     	String proxyPortSt = System.getProperty("http.proxyPort");
@@ -134,34 +136,34 @@ public class OwnCloudClient extends HttpClient {
     		clearCredentials();
     	}
     }
-    
+
     public void clearCredentials() {
 		if (!(credentials instanceof OwnCloudAnonymousCredentials)) {
 			credentials = OwnCloudCredentialsFactory.getAnonymousCredentials();
 		}
 		credentials.applyTo(this);
 	}
-    
+
     /**
      * Requests the received method with the received timeout (milliseconds).
-     * 
+     * <p>
      * Executes the method through the inherited HttpClient.executedMethod(method).
-     * 
+     * <p>
      * Sets the socket and connection timeouts only for the method received.
-     * 
-     * The timeouts are both in milliseconds; 0 means 'infinite'; 
+     * <p>
+     * The timeouts are both in milliseconds; 0 means 'infinite';
      * < 0 means 'do not change the default'
      *
-     * @param method                HTTP method request.
-     * @param readTimeout           Timeout to set for data reception
-     * @param connectionTimeout     Timeout to set for connection establishment
+     * @param method            HTTP method request.
+     * @param readTimeout       Timeout to set for data reception
+     * @param connectionTimeout Timeout to set for connection establishment
      */
     public int executeMethod(HttpMethodBase method, int readTimeout, int connectionTimeout) throws IOException {
 
         int oldSoTimeout = getParams().getSoTimeout();
         int oldConnectionTimeout = getHttpConnectionManager().getParams().getConnectionTimeout();
         try {
-            if (readTimeout >= 0) { 
+            if (readTimeout >= 0) {
                 method.getParams().setSoTimeout(readTimeout);   // this should be enough...
                 getParams().setSoTimeout(readTimeout);          // ... but HTTPS needs this
             }
@@ -185,6 +187,8 @@ public class OwnCloudClient extends HttpClient {
      */
     @Override
     public int executeMethod(HttpMethod method) throws IOException {
+        final String hostname = method.getURI().getHost();
+
         try {
             // Update User Agent
             HttpParams params = method.getParams();
@@ -199,6 +203,10 @@ public class OwnCloudClient extends HttpClient {
 
             int status = super.executeMethod(method);
 
+            if (status >= 500 && status < 600 && DNSCache.INSTANCE.isIPV6(hostname)) {
+                return retryMethodWithIPv4(method, hostname);
+            }
+
             if (followRedirects) {
                 status = followRedirection(method).getLastStatus();
             }
@@ -209,30 +217,41 @@ public class OwnCloudClient extends HttpClient {
 
             return status;
 
+        } catch (SocketTimeoutException e) {
+            if (DNSCache.INSTANCE.isIPV6(hostname)) {
+                return retryMethodWithIPv4(method, hostname);
+            } else {
+                throw e;
+            }
         } catch (IOException e) {
             //Log_OC.d(TAG + " #" + mInstanceNumber, "Exception occurred", e);
             throw e;
         }
     }
 
+    private int retryMethodWithIPv4(HttpMethod method, String hostname) throws IOException {
+        DNSCache.INSTANCE.setIPVersionPreference(hostname, true);
+        return executeMethod(method);
+    }
 
-	public RedirectionPath followRedirection(HttpMethod method) throws IOException {
+
+    public RedirectionPath followRedirection(HttpMethod method) throws IOException {
         int redirectionsCount = 0;
         int status = method.getStatusCode();
         RedirectionPath result = new RedirectionPath(status, MAX_REDIRECTIONS_COUNT);
         while (redirectionsCount < MAX_REDIRECTIONS_COUNT &&
-                (   status == HttpStatus.SC_MOVED_PERMANENTLY || 
-                    status == HttpStatus.SC_MOVED_TEMPORARILY ||
-                    status == HttpStatus.SC_TEMPORARY_REDIRECT)
+                (status == HttpStatus.SC_MOVED_PERMANENTLY ||
+                        status == HttpStatus.SC_MOVED_TEMPORARILY ||
+                        status == HttpStatus.SC_TEMPORARY_REDIRECT)
                 ) {
-            
+
             Header location = method.getResponseHeader("Location");
             if (location == null) {
             	location = method.getResponseHeader("location");
             }
             if (location != null) {
-                Log_OC.d(TAG + " #" + mInstanceNumber,  
-                		"Location to redirect: " + location.getValue());
+                Log_OC.d(TAG + " #" + mInstanceNumber,
+                        "Location to redirect: " + location.getValue());
 
                 String locationStr = location.getValue();
                 result.addLocation(locationStr);
@@ -252,16 +271,16 @@ public class OwnCloudClient extends HttpClient {
                     String redirectionBase = locationStr.substring(0, suffixIndex);
 
                     String destinationStr = destination.getValue();
-                	String destinationPath = destinationStr.substring(baseUri.toString().length());
-                	String redirectedDestination = redirectionBase + destinationPath;
-                	
-                	destination.setValue(redirectedDestination);
+                    String destinationPath = destinationStr.substring(baseUri.toString().length());
+                    String redirectedDestination = redirectionBase + destinationPath;
+
+                    destination.setValue(redirectedDestination);
                     method.setRequestHeader(destination);
                 }
                 status = super.executeMethod(method);
                 result.addStatus(status);
                 redirectionsCount++;
-                
+
             } else {
                 Log_OC.d(TAG + " #" + mInstanceNumber,  "No location to redirect!");
                 status = HttpStatus.SC_NOT_FOUND;
@@ -272,7 +291,7 @@ public class OwnCloudClient extends HttpClient {
 
 	/**
      * Exhausts a not interesting HTTP response. Encouraged by HttpClient documentation.
-     * 
+     *
      * @param responseBodyAsStream      InputStream with the HTTP response to exhaust.
      */
     public void exhaustResponse(InputStream responseBodyAsStream) {
@@ -280,7 +299,7 @@ public class OwnCloudClient extends HttpClient {
             try {
                 while (responseBodyAsStream.read(sExhaustBuffer) >= 0);
                 responseBodyAsStream.close();
-            
+
             } catch (IOException io) {
                 Log_OC.e(TAG, "Unexpected exception while exhausting not interesting HTTP response;" +
                 		" will be IGNORED", io);
@@ -289,7 +308,7 @@ public class OwnCloudClient extends HttpClient {
     }
 
     /**
-     * Sets the connection and wait-for-data timeouts to be applied by default to the methods 
+     * Sets the connection and wait-for-data timeouts to be applied by default to the methods
      * performed by this client.
      */
     public void setDefaultTimeouts(int defaultDataTimeout, int defaultConnectionTimeout) {
@@ -300,7 +319,7 @@ public class OwnCloudClient extends HttpClient {
             getHttpConnectionManager().getParams().setConnectionTimeout(defaultConnectionTimeout);
         }
     }
-    
+
     public String getFilesDavUri(String path) {
         return getDavUri() + "/files/" + getUserId() + "/" + WebdavUtils.encodePath(path);
     }
@@ -367,7 +386,7 @@ public class OwnCloudClient extends HttpClient {
         }
 	}
 
-	private void logSetCookiesAtResponse(Header[] headers) {
+    private void logSetCookiesAtResponse(Header[] headers) {
         int counter = 0;
         for (Header header : headers) {
             if ("set-cookie".equals(header.getName().toLowerCase(Locale.US))) {
@@ -375,42 +394,42 @@ public class OwnCloudClient extends HttpClient {
             }
         }
         if (counter == 0) {
-    		Log_OC.d(TAG + " #" + mInstanceNumber, "No set-cookie");
+            Log_OC.d(TAG + " #" + mInstanceNumber, "No set-cookie");
         }
-        
-	}
-	
-	public String getCookiesString() {
-		Cookie[] cookies = getState().getCookies();
-		String cookiesString = "";
-		for (Cookie cookie : cookies) {
-			cookiesString = cookiesString + cookie.toString() + ";";
 
-			// logCookie(cookie);
-		}
+    }
 
-		return cookiesString;
+    public String getCookiesString() {
+        Cookie[] cookies = getState().getCookies();
+        String cookiesString = "";
+        for (Cookie cookie : cookies) {
+            cookiesString = cookiesString + cookie.toString() + ";";
 
-	}
+            // logCookie(cookie);
+        }
 
-	public int getConnectionTimeout() {
+        return cookiesString;
+
+    }
+
+    public int getConnectionTimeout() {
         return getHttpConnectionManager().getParams().getConnectionTimeout();
-	}
-	
-	public int getDataTimeout() {
-		return getParams().getSoTimeout();
-	}
-	
-	private void logCookie(Cookie cookie) {
-    	Log_OC.d(TAG, "Cookie name: "+ cookie.getName() );
-    	Log_OC.d(TAG, "       value: "+ cookie.getValue() );
-    	Log_OC.d(TAG, "       domain: "+ cookie.getDomain());
-    	Log_OC.d(TAG, "       path: "+ cookie.getPath() );
-    	Log_OC.d(TAG, "       version: "+ cookie.getVersion() );
-    	Log_OC.d(TAG, "       expiryDate: " + 
-    			(cookie.getExpiryDate() != null ? cookie.getExpiryDate().toString() : "--"));
-    	Log_OC.d(TAG, "       comment: "+ cookie.getComment() );
-    	Log_OC.d(TAG, "       secure: "+ cookie.getSecure() );
+    }
+
+    public int getDataTimeout() {
+        return getParams().getSoTimeout();
+    }
+
+    private void logCookie(Cookie cookie) {
+        Log_OC.d(TAG, "Cookie name: " + cookie.getName());
+        Log_OC.d(TAG, "       value: " + cookie.getValue());
+        Log_OC.d(TAG, "       domain: " + cookie.getDomain());
+        Log_OC.d(TAG, "       path: " + cookie.getPath());
+        Log_OC.d(TAG, "       version: " + cookie.getVersion());
+        Log_OC.d(TAG, "       expiryDate: " +
+                (cookie.getExpiryDate() != null ? cookie.getExpiryDate().toString() : "--"));
+        Log_OC.d(TAG, "       comment: " + cookie.getComment());
+        Log_OC.d(TAG, "       secure: " + cookie.getSecure());
     }
 
     /**
