+++ src/main/java/com/owncloud/android/lib/common/network/AdvancedSslSocketFactory.java
@@ -24,6 +24,8 @@
 
 package com.owncloud.android.lib.common.network;
 
+import com.nextcloud.common.DNSCache;
+import com.nextcloud.common.IPV6PreferringDNS;
 import com.owncloud.android.lib.common.utils.Log_OC;
 
 import org.apache.commons.httpclient.ConnectTimeoutException;
@@ -187,7 +189,7 @@ public class AdvancedSslSocketFactory implements SecureProtocolSocketFactory {
     }
 
     private InetAddress getInetAddressForHost(String host) throws UnknownHostException {
-        return InetAddress.getByName(host);
+        return DNSCache.INSTANCE.lookup(host).get(0);
     }
 
     /**
